Fixed compilation on IA32, and vastly reduced the number of remote TLB invalidations that have to take place when accessing physical memory on IA32. This is done by temporarily wiring a thread to its CPU when accessing physical memory in certain cases, and then passing a parameter to page_phys_unmap to cause it to not perform an unnecessary remote TLB invalidation.

+++ source/drivers/disk/partition.c
@@ -36,7 +36,7 @@ static int partition_device_read(device_t *device, void *buf, size_t count, offs
 	partition_t *part = device->data;
 	disk_device_t *disk = part->parent;
 
-	if((offset + count) >= (part->size * disk->blksize)) {
+	if((uint64_t)(offset + count) >= (part->size * disk->blksize)) {
 		*bytesp = 0;
 		return 0;
 	}
@@ -55,7 +55,7 @@ static int partition_device_write(device_t *device, const void *buf, size_t coun
 	partition_t *part = device->data;
 	disk_device_t *disk = part->parent;
 
-	if((offset + count) >= (part->size * disk->blksize)) {
+	if((uint64_t)(offset + count) >= (part->size * disk->blksize)) {
 		*bytesp = 0;
 		return 0;
 	}
